more gcloud provisioning updates

+++ provisioning/provision.sh
@@ -1,18 +1,21 @@
+if [ -d /mnt/disks/mimic ]; then
+	HOME=/mnt/disks/mimic
+fi
 CONDA_VER=latest
 OS_TYPE=$(uname -i)
+if [[ "$OS_TYPE" == "unknown" ]]; then
+	OS_TYPE=x86_64
+fi
 CONDA_DIR=${HOME}/miniconda
 PY_VER=3.11
+
 cd ${HOME} 
 
-if [[ "$OS_TYPE=="unknown" ]]; then
-	OS_TYPE="x86_64"
-fi
-
 if [ ! -d "${CONDA_DIR}" ]; then
-	echo "removing local python site-packages..."
-	rm -rf ${HOME}/.local/lib/python${PY_VER}/site-packages
-	echo "installing conda..."
-	curl -LO "http://repo.continuum.io/miniconda/Miniconda3-${CONDA_VER}-Linux-${OS_TYPE}.sh"
+	if [ -d "${HOME}/.local/lib/python${PY_VER}/site-packages" ]; then
+		rm -rf ${HOME}/.local/lib/python${PY_VER}/site-packages
+	fi
+	curl -LO "https://repo.continuum.io/miniconda/Miniconda3-${CONDA_VER}-Linux-${OS_TYPE}.sh"
 	bash Miniconda3-${CONDA_VER}-Linux-${OS_TYPE}.sh -p ${HOME}/miniconda -b
 	rm Miniconda3-${CONDA_VER}-Linux-${OS_TYPE}.sh
 fi
@@ -60,10 +63,8 @@ else
 	conda env create --name mimic --file=git/mimicChess/provisioning/environment.yml
 fi
 
-echo "set -g mouse on" > ${HOME}/.tmux.conf
-
 sudo apt update
-sudo apt install -y -V ca-certificates lsb-release wget stockfish libzstd-dev clangd
+sudo apt install -y -V ca-certificates lsb-release wget stockfish libzstd-dev clangd tmux
 wget https://packages.apache.org/artifactory/arrow/$(lsb_release --id --short | tr 'A-Z' 'a-z')/apache-arrow-apt-source-latest-$(lsb_release --codename --short).deb
 sudo apt install -y -V ./apache-arrow-apt-source-latest-$(lsb_release --codename --short).deb
 sudo apt update
@@ -81,3 +82,5 @@ sudo apt install -y -V libgandiva-glib-dev # For Gandiva GLib (C)
 sudo apt install -y -V libparquet-dev # For Apache Parquet C++
 sudo apt install -y -V libparquet-glib-dev # For Apache Parquet GLib (C)
 rm apache-arrow-apt-source-latest-$(lsb_release --codename --short).deb
+
+echo "set -g mouse on" > ${HOME}/.tmux.conf
